backend/requirements.txt
Flask
pymongo
python-dotenv
pyjwt
flask-cors